Finally replace old package with new one
