--- archiver/testdata/proverbs/extra/proverb3.txt
@@ -1,2 +1,0 @@
-"interface{} says nothing."
-	- Rob Pike
