--- archiver/testdata/proverbs/proverb1.txt
@@ -1,2 +1,0 @@
-"Channels orchestrate; mutexes serialize."
-	- Rob Pike
