--- archiver/testdata/quote1.txt
@@ -1,2 +1,0 @@
-"Go has generics; they're called interfaces."
-	- Matt Holt
