+++ tarsz.go
@@ -3,90 +3,98 @@ package archiver
 import (
 	"fmt"
 	"io"
-	"os"
 	"strings"
 
 	"github.com/golang/snappy"
 )
 
-// TarSz is for TarSz format
-var TarSz tarSzFormat
-
-func init() {
-	RegisterFormat("TarSz", TarSz)
+// TarSz facilitates Snappy compression
+// (https://github.com/google/snappy)
+// of tarball archives.
+type TarSz struct {
+	*Tar
 }
 
-type tarSzFormat struct{}
-
-func (tarSzFormat) Match(filename string) bool {
-	return strings.HasSuffix(strings.ToLower(filename), ".tar.sz") || strings.HasSuffix(strings.ToLower(filename), ".tsz") || isTarSz(filename)
-}
-
-// isTarSz checks the file has the sz compressed Tar format header by
-// reading its beginning block.
-func isTarSz(tarszPath string) bool {
-	f, err := os.Open(tarszPath)
-	if err != nil {
-		return false
+// Archive creates a compressed tar file at destination
+// containing the files listed in sources. The destination
+// must end with ".tar.sz" or ".tsz". File paths can be
+// those of regular files or directories; directories will
+// be recursively added.
+func (tsz *TarSz) Archive(sources []string, destination string) error {
+	if !strings.HasSuffix(destination, ".tar.sz") &&
+		!strings.HasSuffix(destination, ".tsz") {
+		return fmt.Errorf("output filename must have .tar.sz or .tsz extension")
 	}
-	defer f.Close()
+	tsz.wrapWriter()
+	return tsz.Tar.Archive(sources, destination)
+}
 
-	szr := snappy.NewReader(f)
-	buf := make([]byte, tarBlockSize)
-	n, err := szr.Read(buf)
-	if err != nil || n < tarBlockSize {
-		return false
+// Unarchive unpacks the compressed tarball at
+// source to destination. Destination will be
+// treated as a folder name.
+func (tsz *TarSz) Unarchive(source, destination string) error {
+	tsz.wrapReader()
+	return tsz.Tar.Unarchive(source, destination)
+}
+
+// Walk calls walkFn for each visited item in archive.
+func (tsz *TarSz) Walk(archive string, walkFn WalkFunc) error {
+	tsz.wrapReader()
+	return tsz.Tar.Walk(archive, walkFn)
+}
+
+// Create opens tsz for writing a compressed
+// tar archive to out.
+func (tsz *TarSz) Create(out io.Writer) error {
+	tsz.wrapWriter()
+	return tsz.Create(out)
+}
+
+// Open opens t for reading a compressed archive from
+// in. The size parameter is not used.
+func (tsz *TarSz) Open(in io.Reader, size int64) error {
+	tsz.wrapReader()
+	return tsz.Tar.Open(in, size)
+}
+
+// Extract extracts a single file from the tar archive.
+// If the target is a directory, the entire folder will
+// be extracted into destination.
+func (tsz *TarSz) Extract(source, target, destination string) error {
+	tsz.wrapReader()
+	return tsz.Tar.Extract(source, target, destination)
+}
+
+func (tsz *TarSz) wrapWriter() {
+	var sw *snappy.Writer
+	tsz.Tar.writerWrapFn = func(w io.Writer) (io.Writer, error) {
+		sw = snappy.NewWriter(w)
+		return sw, nil
 	}
-
-	return hasTarHeader(buf)
-}
-
-// Write outputs a .tar.sz file to a Writer containing
-// the contents of files listed in filePaths. File paths
-// can be those of regular files or directories. Regular
-// files are stored at the 'root' of the archive, and
-// directories are recursively added.
-func (tarSzFormat) Write(output io.Writer, filePaths []string) error {
-	return writeTarSz(filePaths, output, "")
-}
-
-// Make creates a .tar.sz file at tarszPath containing
-// the contents of files listed in filePaths. File paths
-// can be those of regular files or directories. Regular
-// files are stored at the 'root' of the archive, and
-// directories are recursively added.
-func (tarSzFormat) Make(tarszPath string, filePaths []string) error {
-	out, err := os.Create(tarszPath)
-	if err != nil {
-		return fmt.Errorf("error creating %s: %v", tarszPath, err)
+	tsz.Tar.cleanupWrapFn = func() {
+		sw.Close()
 	}
-	defer out.Close()
-
-	return writeTarSz(filePaths, out, tarszPath)
 }
 
-func writeTarSz(filePaths []string, output io.Writer, dest string) error {
-	szw := snappy.NewBufferedWriter(output)
-	defer szw.Close()
-
-	return writeTar(filePaths, szw, dest)
-}
-
-// Read untars a .tar.sz file read from a Reader and decompresses
-// the contents into destination.
-func (tarSzFormat) Read(input io.Reader, destination string) error {
-	szr := snappy.NewReader(input)
-
-	return Tar.Read(szr, destination)
-}
-
-// Open untars source and decompresses the contents into destination.
-func (tarSzFormat) Open(source, destination string) error {
-	f, err := os.Open(source)
-	if err != nil {
-		return fmt.Errorf("%s: failed to open archive: %v", source, err)
+func (tsz *TarSz) wrapReader() {
+	tsz.Tar.readerWrapFn = func(r io.Reader) (io.Reader, error) {
+		return snappy.NewReader(r), nil
 	}
-	defer f.Close()
+}
 
-	return TarSz.Read(f, destination)
+func (tsz *TarSz) String() string { return "tar.sz" }
+
+// Compile-time checks to ensure type implements desired interfaces.
+var (
+	_ = Reader(new(TarSz))
+	_ = Writer(new(TarSz))
+	_ = Archiver(new(TarSz))
+	_ = Unarchiver(new(TarSz))
+	_ = Walker(new(TarSz))
+	_ = Extractor(new(TarSz))
+)
+
+// DefaultTarSz is a convenient archiver ready to use.
+var DefaultTarSz = &TarSz{
+	Tar: DefaultTar,
 }
