+++ tarxz.go
@@ -3,103 +3,103 @@ package archiver
 import (
 	"fmt"
 	"io"
-	"os"
 	"strings"
 
 	"github.com/ulikunitz/xz"
+	fastxz "github.com/xi2/xz"
 )
 
-// TarXZ is for TarXZ format
-var TarXZ xzFormat
-
-func init() {
-	RegisterFormat("TarXZ", TarXZ)
+// TarXz facilitates xz compression
+// (https://tukaani.org/xz/format.html)
+// of tarball archives.
+type TarXz struct {
+	*Tar
 }
 
-type xzFormat struct{}
-
-// Match returns whether filename matches this format.
-func (xzFormat) Match(filename string) bool {
-	return strings.HasSuffix(strings.ToLower(filename), ".tar.xz") ||
-		strings.HasSuffix(strings.ToLower(filename), ".txz") ||
-		isTarXz(filename)
+// Archive creates a compressed tar file at destination
+// containing the files listed in sources. The destination
+// must end with ".tar.gz" or ".txz". File paths can be
+// those of regular files or directories; directories will
+// be recursively added.
+func (txz *TarXz) Archive(sources []string, destination string) error {
+	if !strings.HasSuffix(destination, ".tar.xz") &&
+		!strings.HasSuffix(destination, ".txz") {
+		return fmt.Errorf("output filename must have .tar.xz or .txz extension")
+	}
+	txz.wrapWriter()
+	return txz.Tar.Archive(sources, destination)
 }
 
-// isTarXz checks the file has the xz compressed Tar format header by reading
-// its beginning block.
-func isTarXz(tarxzPath string) bool {
-	f, err := os.Open(tarxzPath)
-	if err != nil {
-		return false
-	}
-	defer f.Close()
-
-	xzr, err := xz.NewReader(f)
-	if err != nil {
-		return false
-	}
-
-	buf := make([]byte, tarBlockSize)
-	n, err := xzr.Read(buf)
-	if err != nil || n < tarBlockSize {
-		return false
-	}
-
-	return hasTarHeader(buf)
+// Unarchive unpacks the compressed tarball at
+// source to destination. Destination will be
+// treated as a folder name.
+func (txz *TarXz) Unarchive(source, destination string) error {
+	txz.wrapReader()
+	return txz.Tar.Unarchive(source, destination)
 }
 
-// Write outputs a .tar.xz file to a Writer containing
-// the contents of files listed in filePaths. File paths
-// can be those of regular files or directories. Regular
-// files are stored at the 'root' of the archive, and
-// directories are recursively added.
-func (xzFormat) Write(output io.Writer, filePaths []string) error {
-	return writeTarXZ(filePaths, output, "")
+// Walk calls walkFn for each visited item in archive.
+func (txz *TarXz) Walk(archive string, walkFn WalkFunc) error {
+	txz.wrapReader()
+	return txz.Tar.Walk(archive, walkFn)
 }
 
-// Make creates a .tar.xz file at xzPath containing
-// the contents of files listed in filePaths. File
-// paths can be those of regular files or directories.
-// Regular files are stored at the 'root' of the
-// archive, and directories are recursively added.
-func (xzFormat) Make(xzPath string, filePaths []string) error {
-	out, err := os.Create(xzPath)
-	if err != nil {
-		return fmt.Errorf("error creating %s: %v", xzPath, err)
-	}
-	defer out.Close()
-
-	return writeTarXZ(filePaths, out, xzPath)
+// Create opens txz for writing a compressed
+// tar archive to out.
+func (txz *TarXz) Create(out io.Writer) error {
+	txz.wrapWriter()
+	return txz.Create(out)
 }
 
-func writeTarXZ(filePaths []string, output io.Writer, dest string) error {
-	xzw, err := xz.NewWriter(output)
-	if err != nil {
-		return fmt.Errorf("error compressing xz: %v", err)
-	}
-	defer xzw.Close()
-
-	return writeTar(filePaths, xzw, dest)
+// Open opens t for reading a compressed archive from
+// in. The size parameter is not used.
+func (txz *TarXz) Open(in io.Reader, size int64) error {
+	txz.wrapReader()
+	return txz.Tar.Open(in, size)
 }
 
-// Read untars a .tar.xz file read from a Reader and decompresses
-// the contents into destination.
-func (xzFormat) Read(input io.Reader, destination string) error {
-	xzr, err := xz.NewReader(input)
-	if err != nil {
-		return fmt.Errorf("error decompressing xz: %v", err)
-	}
-
-	return Tar.Read(xzr, destination)
+// Extract extracts a single file from the tar archive.
+// If the target is a directory, the entire folder will
+// be extracted into destination.
+func (txz *TarXz) Extract(source, target, destination string) error {
+	txz.wrapReader()
+	return txz.Tar.Extract(source, target, destination)
 }
 
-// Open untars source and decompresses the contents into destination.
-func (xzFormat) Open(source, destination string) error {
-	f, err := os.Open(source)
-	if err != nil {
-		return fmt.Errorf("%s: failed to open archive: %v", source, err)
+func (txz *TarXz) wrapWriter() {
+	var xzw *xz.Writer
+	txz.Tar.writerWrapFn = func(w io.Writer) (io.Writer, error) {
+		var err error
+		xzw, err = xz.NewWriter(w)
+		return xzw, err
 	}
-	defer f.Close()
+	txz.Tar.cleanupWrapFn = func() {
+		xzw.Close()
+	}
+}
 
-	return TarXZ.Read(f, destination)
+func (txz *TarXz) wrapReader() {
+	var xzr *fastxz.Reader
+	txz.Tar.readerWrapFn = func(r io.Reader) (io.Reader, error) {
+		var err error
+		xzr, err = fastxz.NewReader(r, 0)
+		return xzr, err
+	}
+}
+
+func (txz *TarXz) String() string { return "tar.xz" }
+
+// Compile-time checks to ensure type implements desired interfaces.
+var (
+	_ = Reader(new(TarXz))
+	_ = Writer(new(TarXz))
+	_ = Archiver(new(TarXz))
+	_ = Unarchiver(new(TarXz))
+	_ = Walker(new(TarXz))
+	_ = Extractor(new(TarXz))
+)
+
+// DefaultTarXz is a convenient archiver ready to use.
+var DefaultTarXz = &TarXz{
+	Tar: DefaultTar,
 }
